Remove extraneous files from fixtures and update lock files
Many of the fixtures included files that were not actually relevant to
any tests in this repo and could be removed (with only one actual test
modification to write to a different path).

There were also some placeholder package names ("package-2") as deps
which actually happen to exist, so those were change to "is-number"
which is well-known and has no deps.

Finally, some of the lock files were outdated/placeholders and were
updated to reduce confusion in the lock file refreshing PRs.