Update index.js
In this modified code, I've added a try-catch block to catch any errors that may occur during the execution of the ExtractorConfig.prepare and Extractor.invoke functions. If an error occurs, it will log the error message and stack trace and exit the process with a status code of 1. Additionally, it generates a bug report containing the error details, which you can save to a file or send to a bug tracking system for further investigation.